requirements.txt
asgiref==3.5.2
awscli==1.27.32
botocore==1.29.32
certifi == 2022.9.24
colorama==0.4.4
dj-database-url==1.2.0
Django==4.1
django-environ==0.9.0
django-heroku==0.3.1
docutils==0.16
gunicorn==20.1.0
jmespath==1.0.1
psycopg2==2.9.5
pyasn1==0.4.8
python-dateutil==2.8.2
python-decouple==3.6
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
wincertstore==0.2
